all: handle type aliases
We only ever convert to primitives for less than comparisons. Type
aliases throw a wrench in this because we cannot just type convert from
interface{} to a primitive. We can easily solve this in lteq because
reflect.Value provides handy Int, Uint, Float, and Complex functions.

Unfortunately, for sorting, it would be beneficial for us to directly
convert to the slice of primitives. This is trickier. We do this by
creating a typed slice and then using the defined unsafe / SliceHeader
rules to set the typed slice directly. This circumvents our type alias
problem.