sort: sort innards before sorting the outer type
Sorting struct innards ensures that when we go to sort a slice of the
outer struct itself, we sort properly. If we saved the innards
processing for last, we could inadvertently end up with outer structs
out of order.

This adds a test for the behavior, which also covers testing the prior
commit.